Inventory now functions well and Allies can pickup and shoot weapons
This is a great first step to my ai framework. I'm thinking about adding
some basic functionality in the future. For instance, the line of sight
raycast needs to determine whether an ally shoots or not. Also, if an
ally runs out of ammo, then they either need to reload their gun or find
another one.